% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adae}
\alias{adae}
\title{Adverse Events Dataset.}
\format{
Data frame with 1934 features and 92 fields
}
\source{
\url{https://https://github.com/insightsengineering/random.cdisc.data/tree/main/data}
}
\usage{
data(adae)
}
\description{
Adverse Events Dataset.
}
\seealso{
Other datasets: 
\code{\link{adlb}},
\code{\link{adsl}},
\code{\link{advs}}
}
\concept{datasets}
\keyword{datasets}
