% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adae_by_sev_tox.R
\name{adae_by_sev_tox}
\alias{adae_by_sev_tox}
\title{Adverse Events Table by Body System and Severity/Toxicity}
\usage{
adae_by_sev_tox(
  adsl,
  df_adae,
  colsby = "ARM",
  grade_val = "AESEV",
  class_val = "AESOC",
  term_val = "AEDECOD",
  default_view = TRUE
)
}
\arguments{
\item{adsl}{adsl data}

\item{df_adae}{adae data}

\item{colsby}{split columns by (default: ARM)}

\item{grade_val}{AE Severity or AE Toxicity Grade}

\item{class_val}{Sytem Organ Class/Body System Class}

\item{term_val}{Dictionary Derived Term}

\item{default_view}{Logical `TRUE` or `FALSE`}
}
\value{
An rtable object
}
\description{
Adverse Events Table by Body System and Severity/Toxicity
}
\examples{

adsl <- random.cdisc.data::cadsl
adae <- random.cdisc.data::cadae

ae_by_sev_tox_table <- adae_by_sev_tox(
  adsl = adsl,
  df_adae = adae,
  colsby = "ARM",
  grade_val = "AESEV",
  class_val = "AESOC",
  term_val = "AEDECOD",
  default_view = TRUE
)

}
\seealso{
Other helpers: 
\code{\link{build_adsl}()}
}
\concept{helpers}
\keyword{helpers}
