% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adae_utils.R
\name{add_adae_flags}
\alias{add_adae_flags}
\title{Add Flags to ADAE}
\usage{
add_adae_flags(df)
}
\arguments{
\item{df}{(\code{data.frame})\cr ADAE dataset.}
}
\value{
\code{ADAE} dataset with added flags
}
\description{
Add Flags to ADAE
}
\examples{
adsl <- random.cdisc.data::cadsl
adae <- random.cdisc.data::cadae
adae_ <- add_adae_flags(adae)

tbl <- dplyr::select(adae_, setdiff(names(adae_), names(adae)))
reactable::reactable(dplyr::slice_head(tbl, n = 10))

}
\seealso{
Other adae_utils: 
\code{\link{build_adae_by_sev_tox}()},
\code{\link{build_adae_summary}()}
}
\concept{adae_utils}
\keyword{adae_utils}
