% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adlb}
\alias{adlb}
\title{Laboratory Analysis Dataset.}
\format{
Data frame with 8400 features and 102 fields
}
\source{
\url{https://https://github.com/insightsengineering/random.cdisc.data/tree/main/data}
}
\usage{
data(adlb)
}
\description{
Laboratory Analysis Dataset.
}
\seealso{
Other datasets: 
\code{\link{adae}},
\code{\link{adsl}},
\code{\link{advs}}
}
\concept{datasets}
\keyword{datasets}
