% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adsl}
\alias{adsl}
\title{Subject Level Analysis Dataset.}
\format{
Data frame with 400 features and 55 fields
}
\source{
\url{https://https://github.com/insightsengineering/random.cdisc.data/tree/main/data}
}
\usage{
data(adsl)
}
\description{
Subject Level Analysis Dataset.
}
\seealso{
Other datasets: 
\code{\link{adae}},
\code{\link{adlb}},
\code{\link{advs}}
}
\concept{datasets}
\keyword{datasets}
