% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adae_utils.R
\name{build_adae_summary}
\alias{build_adae_summary}
\title{Create ADAE Summary Table}
\usage{
build_adae_summary(adae, filter_cond = NULL, event_vars, trt_var)
}
\arguments{
\item{adae}{Input \code{ADAE} dataset}

\item{filter_cond}{filter condition}

\item{event_vars}{Variables added to source \code{ADAE} by \code{add_adae_flags()}}

\item{trt_var}{Treatment Variable eg. \code{ARM}}
}
\value{
List containing layout object of ADAE Summary Table and filtered ADAE data
}
\description{
Create ADAE Summary Table
}
\examples{

library(rtables)
adsl <- random.cdisc.data::cadsl
adae <- random.cdisc.data::cadae
adae_ <- add_adae_flags(adae)
lyt <- build_adae_summary(
  adae = adae_,
  filter_cond = NULL,
  event_vars = setdiff(names(adae_), names(adae)),
  trt_var = "ARM"
)
build_table(lyt = lyt$lyt, df = lyt$df_out, alt_counts_df = adsl)

}
\seealso{
Other adae_utils: 
\code{\link{add_adae_flags}()},
\code{\link{build_adae_by_sev_tox}()}
}
\concept{adae_utils}
\keyword{adae_utils}
