% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_adsl_display.R
\name{build_adsl}
\alias{build_adsl}
\title{Build adsl dataset}
\usage{
build_adsl(
  title = "x.x: Study Subject Data",
  subtitle = c("x.x.x: Demographic Characteristics",
    "Table x.x.x.x: Demographic Characteristics - Full Analysis Set"),
  footer = "Source: ADSL DDMMYYYY hh:mm; Listing x.xx; SDTM package: DDMMYYYY",
  split_cols_by = "ARM",
  split_rows_by = NULL,
  summ_vars = c("AGE", "SEX", "COUNTRY")
)
}
\arguments{
\item{title}{Title of the demographic table}

\item{subtitle}{Subtitle of the demographic table}

\item{footer}{Footer of the demographic table}

\item{split_cols_by}{Variable to split the table by columns}

\item{split_rows_by}{Variable to split the table by rows}

\item{summ_vars}{Summary variables}
}
\value{
A tabletree object.
}
\description{
A function to build the demographic table full set
}
\examples{
lyt <- build_adsl(
  split_cols_by = "ARM",
  split_rows_by = "SEX",
  summ_vars = c("AGE", "RACE")
)

## Use the tabletree object to within build_table
\dontrun{
rtables::build_table(lyt, random.cdisc.data::cadsl)
}

}
\seealso{
Other helpers: 
\code{\link{adae_by_sev_tox}()}
}
\concept{helpers}
\keyword{helpers}
