% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_generic_bds_table}
\alias{build_generic_bds_table}
\title{Create generic BDS Summary table}
\usage{
build_generic_bds_table(
  bds_df,
  filter_cond = NULL,
  param,
  trt_var,
  visit = "AVISIT",
  disp_vars
)
}
\arguments{
\item{bds_df}{(\code{data.frame})\cr BDS dataset (typically ADVS, ADLB etc)}

\item{filter_cond}{(\code{character})\cr Filtering condition required for \code{bds_df}.}

\item{param}{(\code{character})\cr BDS parameter value from \code{PARAM}}

\item{trt_var}{(\code{character})\cr Arm variable used to split table into columns.}

\item{visit}{(\code{character})\cr Visit variable name eg. \code{AVISIT}}

\item{disp_vars}{(\verb{vector of characters})\cr Variables to summarize}
}
\value{
List containing Generic BDS table layout and filtered BDS data
}
\description{
Create generic BDS Summary table
}
\examples{

library(clinTables)
library(rtables)
adsl <- random.cdisc.data::cadsl
advs <- random.cdisc.data::cadvs

lyt <- build_generic_bds_table(advs,
  param = "Diastolic Blood Pressure",
  trt_var = "ARM", visit = "AVISIT",
  disp_vars = c("AVAL", "CHG")
)

tt_to_flextable(build_table(lyt = lyt$lyt, df = lyt$df_out, alt_counts_df = adsl))

}
\seealso{
Other generic: 
\code{\link{build_adsl_chars_table}()},
\code{\link{build_generic_occurrence_table}()}
}
\concept{generic}
\keyword{generic}
