% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_summary_utils.R
\name{build_generic_occurrence_table}
\alias{build_generic_occurrence_table}
\title{Create Generic Occurrence Summary Table}
\usage{
build_generic_occurrence_table(
  occ_df,
  filter_cond = NULL,
  trt_var,
  dataset,
  class_var,
  term_var
)
}
\arguments{
\item{occ_df}{(\code{data.frame})\cr Occurrence dataset (typically ADAE, ADMH etc)}

\item{filter_cond}{(\code{character})\cr Filtering condition required for \code{occ_df}.}

\item{trt_var}{(\code{character})\cr Arm variable used to split table into columns.}

\item{dataset}{(\code{character})\cr Name of the dataset eg. \code{"cadae"}.}

\item{class_var}{(\code{character})\cr Body system organ class variable.}

\item{term_var}{(\code{character})\cr Preferred term variable from \code{occ_df} to include in the table.}
}
\value{
List containing table layout object of Generic Occurrence Table and
filtered Occurrencedata
}
\description{
Create Generic Occurrence Summary Table
}
\examples{
library(rtables)
library(tern)
library(dplyr)

adsl <- pharmaverseadam::adsl |> drop_missing_cols()
adae <- pharmaverseadam::adae |> drop_missing_cols()
adae <- filter(adae, SAFFL == "Y")

lyt <- build_generic_occurrence_table(
  occ_df = adae,
  filter_cond = NULL,
  trt_var = "ARM",
  dataset = "adae",
  class_var = "AESOC",
  term_var = "AEDECOD"
)
tbl <- build_table(lyt = lyt$lyt, df = lyt$df_out, alt_counts_df = adsl)

\dontrun{
tt_to_flextable(tbl)
}

}
\seealso{
Other generic: 
\code{\link{build_adsl_chars_table}()},
\code{\link{build_disp_table}()},
\code{\link{build_generic_bds_table}()},
\code{\link{build_shift_table}()}
}
\concept{generic}
\keyword{generic}
