% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_shift_table.R
\name{build_shift_table}
\alias{build_shift_table}
\title{Shift Table}
\usage{
build_shift_table(
  adsl,
  bds_df,
  filter_cond = NULL,
  trt_var,
  trt_label = NULL,
  group_var = NULL,
  group_label = NULL
)
}
\arguments{
\item{adsl}{(\code{data.frame})\cr \code{adsl} data set.}

\item{bds_df}{(\code{data.frame})\cr \code{bds} data set eg. \code{adlb}, \code{advs}.}

\item{filter_cond}{(\code{character})\cr Filtering condition required for \code{bds_df}.}

\item{trt_var}{(\code{character})\cr  Name of the treatment variable from \code{adsl} for treatment
count, eg. \code{ARM}.}

\item{trt_label}{(\verb{named vector of character})\cr Label of \code{trt_var}.}

\item{group_var}{(\verb{vector of characters})\cr Names of variables for grouping in addition to
parameter. Default is \code{NULL}.}

\item{group_label}{(\verb{named vector of character})\cr Label of \code{group_var}.}
}
\value{
A Flextable object of the Shift Table
}
\description{
Shift Table
}
\examples{
library(clinTables)
data(adsl)
data(adlb)

build_shift_table(
  adsl = adsl,
  bds_df = adlb,
  trt_var = "ARM",
  trt_label = c(ARM = "Description of Planned Arm"),
  group_var = "AVISIT",
  group_label = c(AVISIT = "Analysis Visit")
)

}
