% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinTables-package.R
\docType{package}
\name{clinTables-package}
\alias{clinTables}
\alias{clinTables-package}
\title{clinTables: Clinical Trial Tables in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Shiny Application for visualizing some of the standard tables used in Clinical Trials.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ahasoplakus.github.io/clinTables/}
  \item \url{https://github.com/ahasoplakus/clinTables/}
  \item Report bugs at \url{https://github.com/ahasoplakus/clinTables/issues/}
}

}
\author{
\strong{Maintainer}: Sukalpo Saha \email{sukalposaha@gmail.com}

}
\keyword{internal}
