% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{drop_missing_cols}
\alias{drop_missing_cols}
\title{Drop Columns which has only missing values}
\usage{
drop_missing_cols(df)
}
\arguments{
\item{df}{data frame}
}
\value{
list of data frames
}
\description{
Drop Columns which has only missing values
}
\examples{

drop_missing_cols(pharmaverseadam::advs)

}
\keyword{internal}
