% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmtables-package.R
\docType{package}
\name{pharmtables-package}
\alias{pharmtables}
\alias{pharmtables-package}
\title{pharmtables: Clinical Trial Tables in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Shiny Application for visualizing some of the standard tables used in Clinical Trials.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ahasoplakus.github.io/pharmtables/}
  \item \url{https://github.com/ahasoplakus/pharmtables/}
  \item Report bugs at \url{https://github.com/ahasoplakus/pharmtables/issues/}
}

}
\author{
\strong{Maintainer}: Sukalpo Saha \email{sukalposaha@gmail.com}

}
\keyword{internal}
