% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adae_utils.R
\name{add_adae_flags}
\alias{add_adae_flags}
\title{Add Flags to ADAE}
\usage{
add_adae_flags(df)
}
\arguments{
\item{df}{(\code{data.frame})\cr ADAE dataset.}
}
\value{
\code{ADAE} dataset with added flags
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
\examples{
suppressMessages(library(dplyr))

adsl <- pharmaverseadam::adsl
adae <- pharmaverseadam::adae

adae_ <- suppressMessages(add_adae_flags(adae))

adae_ |>
  select(c("USUBJID", setdiff(names(adae_), names(adae)))) |>
  slice_head(n = 5)

}
\seealso{
Other adae_utils: 
\code{\link{build_adae_by_sev_tox}()},
\code{\link{build_adae_summary}()}
}
\concept{adae_utils}
\keyword{adae_utils}
