% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adae_utils.R
\name{build_adae_by_sev_tox}
\alias{build_adae_by_sev_tox}
\title{Adverse Events Table by Body System and Severity/Toxicity}
\usage{
build_adae_by_sev_tox(
  adsl,
  df_adae,
  colsby = "ARM",
  filter_cond = NULL,
  grade_val = "AESEV",
  class_val = "AESOC",
  term_val = "AEDECOD",
  default_view = TRUE
)
}
\arguments{
\item{adsl}{(\code{data.frame})\cr ADSL dataset.}

\item{df_adae}{(\code{data.frame})\cr ADAE dataset.}

\item{colsby}{(\code{character})\cr Variable used to split table into columns.}

\item{filter_cond}{(\code{character})\cr Filtering condition required for \code{df_adae}.}

\item{grade_val}{(\code{character})\cr AE Severity or AE Toxicity Grade i.e. \code{AESEV} or \code{AETOXGR}}

\item{class_val}{(\code{character})\cr System Organ Class/Body System Class i.e. \code{AESOC} or \code{AEBODSYS}}

\item{term_val}{(\code{character})\cr Dictionary Derived Term i.e. \code{AEDECOD} or \code{AETERM}}

\item{default_view}{(\code{logical})\cr \code{TRUE} displays values of \code{grade_val} into columns.}
}
\value{
An rtable object
}
\description{
Adverse Events Table by Body System and Severity/Toxicity
}
\examples{

adsl <- random.cdisc.data::cadsl
adae <- random.cdisc.data::cadae

tbl <- build_adae_by_sev_tox(
  adsl = adsl,
  df_adae = adae,
  colsby = "ARM",
  grade_val = "AESEV",
  class_val = "AESOC",
  term_val = "AEDECOD",
  default_view = TRUE
)

tbl

}
\seealso{
Other adae_utils: 
\code{\link{add_adae_flags}()},
\code{\link{build_adae_summary}()}
}
\concept{adae_utils}
\keyword{adae_utils}
