% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adae_utils.R
\name{build_adae_by_sev_tox}
\alias{build_adae_by_sev_tox}
\title{Adverse Events Table by Body System and Severity/Toxicity}
\usage{
build_adae_by_sev_tox(
  adsl,
  df_adae,
  colsby = "ARM",
  grade_val = "AESEV",
  class_val = "AESOC",
  term_val = "AEDECOD",
  default_view = TRUE
)
}
\arguments{
\item{adsl}{Input \code{ADSL} data}

\item{df_adae}{Input \code{ADAE} data}

\item{colsby}{Column Variable (default: \code{ARM})}

\item{grade_val}{AE Severity or AE Toxicity Grade i.e. \code{AESEV} or \code{AETOXGR}}

\item{class_val}{Sytem Organ Class/Body System Class i.e. \code{AESOC} or \code{AEBODSYS}}

\item{term_val}{Dictionary Derived Term i.e. \code{AEDECOD} or \code{AETERM}}

\item{default_view}{Logical \code{TRUE} or \code{FALSE}}
}
\value{
An rtable object
}
\description{
Adverse Events Table by Body System and Severity/Toxicity
}
\examples{

adsl <- random.cdisc.data::cadsl
adae <- random.cdisc.data::cadae

build_adae_by_sev_tox(
  adsl = adsl,
  df_adae = adae,
  colsby = "ARM",
  grade_val = "AESEV",
  class_val = "AESOC",
  term_val = "AEDECOD",
  default_view = TRUE
)


}
\seealso{
Other adae_utils: 
\code{\link{add_adae_flags}()},
\code{\link{build_adae_summary}()}
}
\concept{adae_utils}
\keyword{adae_utils}
