% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_shift_table.R
\name{build_shift_table}
\alias{build_shift_table}
\title{Create Shift Table}
\usage{
build_shift_table(
  adsl,
  bds_df,
  filter_cond = NULL,
  trt_var,
  trt_label = NULL,
  group_var = NULL,
  group_label = NULL,
  default_view = TRUE
)
}
\arguments{
\item{adsl}{(\code{data.frame})\cr \code{adsl} data frame.}

\item{bds_df}{(\code{data.frame})\cr \code{bds} data frame eg. \code{adlb}, \code{advs}.}

\item{filter_cond}{(\code{character})\cr Filtering condition required for \code{bds_df}.}

\item{trt_var}{(\code{character})\cr  Name of the treatment variable from \code{adsl} for treatment
count, eg. \code{ARM}. Must be a vector of length \code{1}.}

\item{trt_label}{(\verb{named vector of character})\cr Label of \code{trt_var}. Must be a \strong{named} vector
of length \code{1}.}

\item{group_var}{(\verb{vector of characters})\cr Names of variables for grouping in addition to
parameter. Default is \code{NULL}.}

\item{group_label}{(\verb{named vector of characters})\cr Label of \code{group_var}.}

\item{default_view}{(\code{logical})\cr If \code{TRUE} (\strong{default}), values of \code{trt_var} are shown in
rows, else in columns.}
}
\value{
A Flextable object of the Shift Table
}
\description{
Shift Tables show the progression of change from the baseline, with the progression often
being along time; the number of subjects is displayed in different range
(e.g. low, normal, or high) at baseline and at selected time points or intervals.
}
\examples{
adsl <- pharmaverseadam::adsl |> drop_missing_cols()
adlb <- pharmaverseadam::adlb |> drop_missing_cols()

\dontrun{
build_shift_table(
  adsl = adsl,
  bds_df = adlb,
  filter_cond = NULL,
  trt_var = "ARM",
  trt_label = c(ARM = "Description of Planned Arm"),
  group_var = NULL, ,
  default_view = FALSE
)
}

}
\seealso{
Other generic: 
\code{\link{build_adsl_chars_table}()},
\code{\link{build_disp_table}()},
\code{\link{build_generic_bds_table}()},
\code{\link{build_generic_occurrence_table}()}
}
\concept{generic}
\keyword{generic}
