% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{filters_to_cond}
\alias{filters_to_cond}
\title{Create filtering condition based on filters}
\usage{
filters_to_cond(filter_list)
}
\arguments{
\item{filter_list}{(\code{list})\cr Named list of filter values}
}
\value{
The Filtering condition
}
\description{
Create filtering condition based on filters
}
\examples{

filter_list <- list(SEX = c("F", "M"))
filters_to_cond(filter_list)

}
\concept{helpers}
\keyword{helpers}
